+++ TabloidFullStack/TabloidFullStack/client/tabloid/src/App.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { BrowserRouter as Router } from "react-router-dom";
 import Header from "./components/Header";
-import ApplicationViews from "./components/Views/ApplicationViews";
+import { ApplicationViews } from "./components/Views/ApplicationViews";
 import { useEffect } from 'react';
 import Authorize from './components/Authorize';
